+++ src/main/scala/org/constellation/domain/redownload/RedownloadService.scala
@@ -7,7 +7,7 @@ import cats.effect.{Concurrent, ContextShift, Timer}
 import cats.implicits._
 import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
 import org.constellation.ConfigUtil
-import org.constellation.checkpoint.CheckpointAcceptanceService
+import org.constellation.checkpoint.{CheckpointAcceptanceService, TopologicalSort}
 import org.constellation.consensus.{FinishedCheckpoint, StoredSnapshot}
 import org.constellation.domain.cloud.HeightHashFileStorage
 import org.constellation.domain.redownload.MajorityStateChooser.SnapshotProposal
@@ -145,10 +145,11 @@ class RedownloadService[F[_]: NonEmptyParallel](
       }
       newProposals = responses.toMap
       proposals <- peersProposals.modify { m =>
-        val updated = m.map { case (id, proposalsAtHeight) =>
-          val diff = newProposals.getOrElse(id, Map.empty) -- proposalsAtHeight.keySet
+        val updated = m.map {
+          case (id, proposalsAtHeight) =>
+            val diff = newProposals.getOrElse(id, Map.empty) -- proposalsAtHeight.keySet
 
-          id -> (proposalsAtHeight ++ diff)
+            id -> (proposalsAtHeight ++ diff)
         } ++ (newProposals -- m.keySet)
 
         val ignored = updated.mapValues(a => takeHighestUntilKey(a, getRemovalPoint(maxHeight(a))))
@@ -278,12 +279,11 @@ class RedownloadService[F[_]: NonEmptyParallel](
           blocksFromSnapshots = acceptedSnapshots.flatMap(_.checkpointCache)
           acceptedBlocksFromSnapshotInfo = snapshotInfoFromMemPool.acceptedCBSinceSnapshotCache
           awaitingBlocksFromSnapshotInfo = snapshotInfoFromMemPool.awaitingCbs
-          blocksToAccept = (blocksFromSnapshots ++ acceptedBlocksFromSnapshotInfo ++ awaitingBlocksFromSnapshotInfo)
-            .sortBy(_.height)
+          blocksToAccept = (blocksFromSnapshots ++ acceptedBlocksFromSnapshotInfo ++ awaitingBlocksFromSnapshotInfo).distinct
 
           _ <- snapshotService.setSnapshot(majoritySnapshotInfo)
 
-          _ <- blocksToAccept.traverse { b =>
+          _ <- TopologicalSort.sortBlocksTopologically(blocksToAccept).traverse { b =>
             logger.debug(s"Accepting block above majority: ${b.height}") >> checkpointAcceptanceService
               .accept(b)
               .handleErrorWith(